+++ pkg/logging/sanitize_handler.go
@@ -3,8 +3,8 @@ package logging
 import "log/slog"
 
 type sanitizeHandler struct {
-	inner          slog.Handler
-	inPrivacyGroup bool
+	inner         slog.Handler
+	inMaskedGroup bool
 }
 
 func newSanitizeHandler(inner slog.Handler) *sanitizeHandler {